Move to ruledef package
