+++ pmd-apex/pom.xml
@@ -39,7 +39,7 @@
         </executions>
         <configuration>
           <includeProjectDependencies>true</includeProjectDependencies>
-          <mainClass>net.sourceforge.pmd.RuleDefsAggregator</mainClass>
+          <mainClass>net.sourceforge.pmd.ruledef.RuleDefsAggregator</mainClass>
           <arguments>
             <argument>${project.basedir}</argument>
             <argument>apex</argument>
