+++ pmd-core/src/main/java/net/sourceforge/pmd/ruledef/RuleDefsAggregator.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd;
+package net.sourceforge.pmd.ruledef;
 
 import java.io.File;
 import java.io.FileInputStream;
