+++ pmd-visualforce/pom.xml
@@ -38,7 +38,7 @@
                 </executions>
                 <configuration>
                     <includeProjectDependencies>true</includeProjectDependencies>
-                    <mainClass>net.sourceforge.pmd.RuleDefsAggregator</mainClass>
+                    <mainClass>net.sourceforge.pmd.ruledef.RuleDefsAggregator</mainClass>
                     <arguments>
                         <argument>${project.basedir}</argument>
                         <argument>vf</argument>
